Refatorando home: Diminuindo complexidade
