+++ src/views/Home/style.scss
@@ -71,7 +71,7 @@
     padding-left: 4.5vw;
 
   }
-  .button {
+  button {
     box-sizing: border-box;
     position: relative;
     margin-left: 14vw;
